+++ service/controller/clusterapi/v29/changedetection/cluster_test.go
@@ -1,4 +1,4 @@
-package detection
+package changedetection
 
 import (
 	"net"
