+++ service/controller/clusterapi/v29/changedetection/error.go
@@ -1,4 +1,4 @@
-package detection
+package changedetection
 
 import (
 	"github.com/giantswarm/microerror"
